Cast the double to int

+++ src/Misc/ArcingProjectileFix.cpp
@@ -11,8 +11,8 @@ DEFINE_HOOK(6FE4F6, TechnoClass_FireAt_ArcingFix, 0)
 		(double)(coordA->X - coordB->X) * (double)(coordA->X - coordB->X) +
 		(double)(coordA->Y - coordB->Y) * (double)(coordA->Y - coordB->Y) +
 		(double)(coordA->Z - coordB->Z) * (double)(coordA->Z - coordB->Z);
-
-	R->EAX(Math::sqrt(distanceSquared));
+	
+	R->EAX((int)Math::sqrt(distanceSquared));
 	
 	return 0x6FE537;
 }
